tier0: ruff burn-down pass 3 — 355 to 249 errors (29.9% reduction)
Pass 3 of the ruff CI burn-down. Combined with passes 1+2,
total reduction is 1845 to 249 (86.5%).

Changes:
- ruff.toml: add T20 ignores for RunPod handler files and agent
  training/evaluation scripts where print is the correct output
  mechanism (diagnostic tools, CLI scripts). Exclude .tmp/ dir.
- F841: auto-fixed 37 unused local variables via --unsafe-fixes.
- RUF100: auto-fixed 2 unused noqa comments.
- SIM118: fixed 1 `key in dict.keys()` -> `key in dict`.
- RUF046: removed 6 unnecessary int() casts.
- RUF005: fixed 8 collection-literal-concatenation ([x] + [y] -> [x, y]).
- B904: added `from None`/`from exc` to 6 raise-in-except blocks.
- RUF015: fixed 3 unnecessary-iterable-allocation (x[0] -> next(iter(x))).

Verification:
- py -3.12 -m pytest: 5245 passed, 116 skipped, 0 failed
- ruff check . --select B904: All checks passed
- No syntax errors introduced

+++ services/quant_foundry/src/quant_foundry/calibration.py
@@ -342,7 +342,7 @@ class Calibrator:
 
             est = LogisticRegression(C=1e10, solver="lbfgs", max_iter=10000)
             X = [[float(p)] for p in raw_probs]
-            y = [int(round(float(v))) for v in labels]
+            y = [round(float(v)) for v in labels]
             # Need at least two classes for logistic regression.
             if len(set(y)) < 2:
                 # Degenerate: all one class. Platt scaling collapses to a
